refactor(milpac): extract getPromotionProgress into milpac-stats
Moves it out of the profile page's private scope so the Discord dossier
card can reach the same next-rank arithmetic, verbatim. Adds coverage
that pins the tier-progress formula, including its actual (previous
rank's threshold, not the current rank's own) starting point.
